+++ src/models/cw721_tx.ts
@@ -96,4 +96,40 @@ export default class CW721Activity extends BaseModel {
       },
     };
   }
+
+  static async getCw721ContractEvents(
+    startBlock: number,
+    endBlock: number,
+    smartContractId?: number,
+    page?: { prevId: number; limit: number }
+  ) {
+    return SmartContractEvent.query()
+      .withGraphFetched('attributes(selectAttribute)')
+      .joinRelated('[message, tx, smart_contract.code]')
+      .where('smart_contract:code.type', 'CW721')
+      .where('tx.height', '>', startBlock)
+      .andWhere('tx.height', '<=', endBlock)
+      .modify((builder) => {
+        if (smartContractId) {
+          builder.andWhere('smart_contract.id', smartContractId);
+        }
+        if (page) {
+          builder
+            .andWhere('smart_contract_event.id', '>', page.prevId)
+            .orderBy('smart_contract_event.id', 'asc')
+            .limit(page.limit);
+        }
+      })
+      .select(
+        'message.sender as sender',
+        'smart_contract.address as contractAddress',
+        'smart_contract_event.action',
+        'smart_contract_event.event_id',
+        'smart_contract_event.index',
+        'smart_contract_event.id as smart_contract_event_id',
+        'tx.hash',
+        'tx.height'
+      )
+      .orderBy('smart_contract_event.id');
+  }
 }
